+++ resources/nodejs.rb
@@ -2,7 +2,7 @@
 # Cookbook:: app_ror
 # Resource:: nodejs
 #
-# Copyright:: 2023, Earth U
+# Copyright:: 2024, Earth U
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -17,92 +17,139 @@
 # limitations under the License.
 
 # Install NodeJS repository for Ubuntu as per:
-# https://github.com/nodesource/distributions#manual-installation
+# https://github.com/nodesource/distributions/wiki/Repository-Manual-Installation
 
 unified_mode true
 
-# NOTE: Always check in https://deb.nodesource.com/node_XX.x/dists/ if
-#       the version is supported on a particular distribution.
+# NOTE: Versions and distribution reference:
+#         https://github.com/nodesource/distributions/blob/master/README.md
 #
-#       If using old nodejs versions no longer supported in the
-#       desired distro as per the official site above, maybe check
-#       the built-in nodejs package of the OS. For example, here are the
-#       built-in versions for specific Ubuntu distros:
-#           node_10.x on focal (20.04)
-#           node_12.x on jammy (22.04)
+#       Version 12.x will use Ubuntu's native repos. There is no longer a
+#       distribution for it in nodesource.
 #
-#       To use the above, simply set the :install_repo property to false,
-#       in which case, the :version attribute will become useless.
-property :version, %w(node_12.x node_14.x node_16.x node_18.x),
-         description: 'LTS version string (node_12.x, node_16.x, etc.)',
+#       Version 14.x will use the old nodesource distribution.
+#
+#       All succeeding versions will use the new 'nodistro' nodesource repo.
+property :version, %w(node_12.x node_14.x node_16.x node_18.x node_20.x),
+         description: 'LTS version string (node_14.x, node_16.x, etc.)',
          name_property: true
 
-property :install_repo, [true, false],
-         description: 'Whether or not to install the Nodesource apt repo',
-         default: true
-
-property :repo_url_prefix, String,
-         description: 'Prefix for the Nodesource apt repo URL string',
-         default: 'https://deb.nodesource.com/'
-
-property :key_url, String,
-         description: 'Remote URL for Nodesource gpg key',
-         default: 'https://deb.nodesource.com/gpgkey/nodesource.gpg.key'
-
-property :keyring, String,
-         description: 'Local keyring for Nodesource gpg key',
-         default: '/usr/share/keyrings/nodesource.gpg'
-
-property :sources_file, String,
-         description: 'Location of apt source file',
-         default: '/etc/apt/sources.list.d/nodesource.list'
-
-# NOTE: Yarn 1.x must be installed globally at first.
-#       Switching to different versions per project (or directory)
-#       can be done with:
-#           yarn set version [version]
+# NOTE: Check latest Yarn major versions here:
+#         https://yarnpkg.com/getting-started/install
+#       Highest possible Yarn versions for each LTS NodeJS version:
+#         node_12.x: Yarn 1.22.22
+#         node_14.x: Yarn 3.8.2
+#         node_16.x: Yarn 3.8.2
+#         node_18.x: Yarn 'stable' (4.2.1)
+#         node_20.x: Yarn 'stable' (4.2.1)
 #
-#       where [version] can be 'classic' (latest 1.x),
-#       'berry' (latest stable 2.x), 'canary' (latest candidate 2.x),
-#       or a specific version number.
+#       Switching Yarn versions in non-corepack yarns:
+#         yarn set version [version]
+#       Switching Yarn versions with corepack enabled:
+#         corepack prepare yarn@[version] --activate
+#
+#       Latest Yarn major versions:
+#         1.22.22
+#         2.4.0
+#         3.8.2
+#         4.2.1
 property :install_yarn, [true, false],
          description: 'Whether to globally install Yarn',
          default: true
 
-action :install do
-  apt_update 'app_ror_nodejs' do
-    ignore_failure true
+property :yarn_version, String,
+         description: 'Defaults to "latest", which installs the latest '\
+                      'possible version of Yarn at the current NodeJS version.',
+         default: 'latest'
+
+property :yarn_user, String,
+         description: 'Main user for this Yarn instance. Default: "ubuntu".',
+         default: 'ubuntu'
+
+action_class do
+  def latest_yarn
+    case new_resource.version
+    when 'node_12.x'
+      '1.22.22'
+    when 'node_14.x', 'node_16.x'
+      '3.8.2'
+    when 'node_18.x', 'node_20.x'
+      'stable'
+    end
   end
+end
 
-  if new_resource.install_repo
-    package 'gpg'
+action :install do
+  # Install NodeJS
 
-    execute 'get_nodesource_key' do
-      command "wget --quiet -O - #{new_resource.key_url} | "\
-              "gpg --dearmor -o #{new_resource.keyring}"
-      not_if  { ::File.exist?(new_resource.keyring) }
-    end
+  if new_resource.version == 'node_12.x'
+    apt_update
 
-    file new_resource.sources_file do
-      content  "deb [signed-by=#{new_resource.keyring}] "\
-               "#{new_resource.repo_url_prefix}#{new_resource.version} "\
-               "#{node['lsb']['codename']} main"
-      notifies :update, 'apt_update[app_ror_nodejs_repo]', :immediately
-    end
-
-    apt_update 'app_ror_nodejs_repo' do
-      action         :nothing
-      ignore_failure true
-    end
-
-    package 'nodejs'
-  else
     package 'nodejs' do
       package_name %w(nodejs npm)
     end
+
+  else
+    fin_key = if new_resource.version == 'node_14.x'
+                'https://deb.nodesource.com/gpgkey/nodesource.gpg.key'
+              else
+                'https://deb.nodesource.com/gpgkey/nodesource-repo.gpg.key'
+              end
+
+    fin_distrib = if new_resource.version == 'node_14.x'
+                    node['lsb']['codename']
+                  else
+                    'nodistro'
+                  end
+
+    add_apt 'nodesource' do
+      keyserver    false
+      key          fin_key
+      key_dearmor  true
+      uri          "https://deb.nodesource.com/#{new_resource.version}"
+      distribution fin_distrib
+      components   ['main']
+    end
+
+    package 'nodejs'
   end
 
+  # Install Yarn
+
   if new_resource.install_yarn
-    include_recipe 'yarn'
+    yver = if new_resource.yarn_version == 'latest'
+             latest_yarn
+           else
+             new_resource.yarn_version
+           end
+
+    if new_resource.version == 'node_12.x'
+      add_apt 'yarn' do
+        keyserver    false
+        key          'https://dl.yarnpkg.com/debian/pubkey.gpg'
+        key_dearmor  true
+        uri          'https://dl.yarnpkg.com/debian/'
+        distribution 'stable'
+        components   ['main']
+      end
+
+      package 'yarn'
+
+      execute "/usr/bin/yarn set version #{yver}" do
+        user new_resource.yarn_user
+        cwd  "/home/#{new_resource.yarn_user}"
+      end
+
+    else
+      execute '/usr/bin/corepack enable' do
+        notifies :run, 'execute[activate_yarn]', :immediately
+      end
+
+      execute 'activate_yarn' do
+        command "/usr/bin/corepack prepare yarn@#{yver} --activate"
+        user    new_resource.yarn_user
+        action  :nothing
+      end
+    end
   end
 end