+++ test/cookbooks/test/recipes/manage_puma.rb
@@ -2,7 +2,7 @@
 # Cookbook:: test
 # Recipe:: manage_puma
 #
-# Copyright:: 2022, Earth U
+# Copyright:: 2024, Earth U
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -16,4 +16,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-app_ror_manage_puma "#{node['test']['base_dir']}/current"
+app_ror_manage_puma "#{node['test']['base_dir']}/current" do
+  unit_type node['test']['puma_unit_type']
+end